requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
mysql-connector-python
python-dotenv